Added implementation for the WolfManager

+++ src/main/java/nl/ramonpeek/managers/WolfManager.java
@@ -40,16 +40,24 @@ public class WolfManager implements IWolfManager {
 
     @Override
     public Wolf deleteWolf(Wolf wolf) {
+        Validator validator = validatorFactory.getValidator();
+        if(wolf == null || !validator.validate(wolf).isEmpty() || !wolfRepo.containsWolf(wolf))
+            return null;
         return wolfRepo.deleteWolf(wolf);
     }
 
     @Override
     public Wolf updateWolf(Wolf requestedWolf, Wolf updatedWolf) {
+        Validator validator = validatorFactory.getValidator();
+        if(requestedWolf == null || updatedWolf == null || !validator.validate(requestedWolf).isEmpty() || !validator.validate(updatedWolf).isEmpty() || !wolfRepo.containsWolf(requestedWolf))
+            return null;
         return wolfRepo.updateWolf(requestedWolf, updatedWolf);
     }
 
     @Override
     public boolean containsWolf(Wolf wolf) {
+        if(wolf == null)
+            return false;
         return wolfRepo.containsWolf(wolf);
     }
 }
